♻️ refactor(monitoring): extract codex progress helpers
Move progress snapshots and inspection summary calculations into a dedicated model module.

Keep the session controller focused on lifecycle and scheduling state.

Validated with type-check, focused monitoring tests, lint, and Vite production build.
